Patch before running prebuild commands.
Sometimes an application provides a script to initialize its build
environment, and it may be required to patch such a script before running
it (typical case is hardcoding developer's local paths). More generally,
"declaritive" ways of changing code (e.g., forceversion, patch, etc.)
should be preferred to prebuild, which should be used as the last resort
(as it is too generic and thus verbose). So, it's right ordering in that
respect either.