common.prepare_source: better detection of errors
when running 'android update', sometimes the program returns an error on
stderr without giving a !=0 errorcode.
This patch detects such an error, and throws a proper BuildException

(in my case, it printed the following to stderr:
Error: The project either has no target set or the target is invalid.
Please provide a --target to the 'android update' command.)