Added dates for when added and last updated to repo index
Under each 'app' element there is now an 'added' and a 'lastupdated'
element with the date the application was originally added to the repo,
and when the most recent version was added. Under each 'apk' element there
is also an 'added' for that particular version. For all three of these,
the element contains just a date in YYYY-MM-DD format.