Print progress messages as a build goes thru various stages.
Without this, we get bunch of messages at the start of build, which
end with "Applying patch", and then all the things hangs for several
minutes (or more, in case of network problems, etc). So, consistently
keep user aware od what's happening. Note that --verbose switch affects
deatiledness and amount of output (in particular, if it's given, "build"
command output goes to console instead of log).