Improved the description of "Auto Name" in Manual
It was not clear, at least to me, where Auto Name is exactly used. This
change clarifies that it's actually used by the auto update
scripts. Furthermore replace the "It", whis "The Auto Name entry", since
it's not obvious if the "It" refers to "Auto Name" or the
"commit (description)".