convert setup.py into working setuptools-based installer
With this setup.py, its possible to install the required packages using:
  pip install -e .

The Debian packaging will also automatically get the dependencies from
install_requires.  This does not handle the generation of the docs at all.
I have not found a straightforward way to include running ./gendocs.sh in
setup.py, but its easy to run in the Debian packaging.