Write buildserver ID to buildserver on creation
The ID of the buildserver is just the corresponding commit hash in
fdroidserver at the point where it was built, which should be enough to
uniquely identify a version of it.

The idea is that this will ultimately end up as metadata in an apk, such
that its precise lineage can be traced, and the exact same binary can be
reproduced.