replace redundant build_tools check that breaks `fdroid init`
This reverts b637568a624cc75f4e67a0df59a017872e2129c6 since it added a
redundant check that broke `fdroid init` when the default version dir of
build_tools does not exist on the local system.  It then uses the function
that was already in place for checking the build_tools setup in a way that
does not break `fdroid init`.

Now that the fake android home version is not matching the default version,
the tests will catch this bug in the future.